+++ infrastructure/repository/room/impl.go
@@ -7,6 +7,7 @@ import (
 	"github.com/oklog/ulid"
 
 	domain "github.com/karamaru-alpha/chat-go-server/domain/model/room"
+	mysql "github.com/karamaru-alpha/chat-go-server/infrastructure/mysql/room"
 )
 
 type repositoryImpl struct {
@@ -22,13 +23,13 @@ func NewRepositoryImpl(db *gorm.DB) domain.IRepository {
 
 // Save トークルームをDBに保存し永続化させる
 func (r repositoryImpl) Save(entity *domain.Room) error {
-	dto := ToDTO(entity)
+	dto := mysql.ToDTO(entity)
 	return r.db.Create(dto).Error
 }
 
 // FindAll トークルーム一覧をDBから再構築する
 func (r repositoryImpl) FindAll() (*[]domain.Room, error) {
-	var dtos []Room
+	var dtos []mysql.Room
 
 	if err := r.db.Find(&dtos).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -37,12 +38,12 @@ func (r repositoryImpl) FindAll() (*[]domain.Room, error) {
 		return nil, err
 	}
 
-	return ToEntities(&dtos)
+	return mysql.ToEntities(&dtos)
 }
 
 // Find トークルームをIDから１件検索・再構成する
 func (r repositoryImpl) Find(id *domain.ID) (*domain.Room, error) {
-	var dto Room
+	var dto mysql.Room
 
 	if err := r.db.Where("id = ?", ulid.ULID(*id).String()).First(&dto).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -51,12 +52,12 @@ func (r repositoryImpl) Find(id *domain.ID) (*domain.Room, error) {
 		return nil, err
 	}
 
-	return ToEntity(&dto)
+	return mysql.ToEntity(&dto)
 }
 
 // FindByTitle トークルームをタイトルから一件取得する
 func (r repositoryImpl) FindByTitle(title *domain.Title) (*domain.Room, error) {
-	var dto Room
+	var dto mysql.Room
 
 	if err := r.db.Where("title = ?", string(*title)).First(&dto).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -65,5 +66,5 @@ func (r repositoryImpl) FindByTitle(title *domain.Title) (*domain.Room, error) {
 		return nil, err
 	}
 
-	return ToEntity(&dto)
+	return mysql.ToEntity(&dto)
 }
